+++ ChangeLog.md
@@ -1,5 +1,6 @@
 #### v1.0.10
 * Fix bug preloading dictionary from another assembly
+* Add name of UID to DICOM Dump elements
 
 #### v1.0.9 (08/02/2012)
 * Parsing of Attribute Tag element type